+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     the same C++ object is passed twice to platform (Java/Swift/Dart) side, it is now guaranteed to
     be the same object on platform side as well.
 ### Bug fixes:
+  * Fixed missing documentation comments on parameters marked with `@Cpp(CString)`.
   * Fixed handling of nested lists in IDL doc comments.
 
 ## 6.6.1